final push
final changes for backend

+++ app/Services/CategoryService.php
@@ -64,6 +64,10 @@ class CategoryService
 
         $trashed = $category->trashed();
 
+        if(!$trashed && $category->progress()->exists()){
+            return null;
+        }
+
         if (!$trashed) {
             $category->delete();
         } else {
